requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.1
django-celery-beat==2.8.1
django-timezone-field==7.1
djangorestframework==3.16.0
kombu==5.5.3
prompt_toolkit==3.0.51
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==6.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
dotenv~=0.9.9
python-dotenv~=1.1.0
drf-spectacular
pytz
requests
psycopg2